webapp/requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
statsd==4.0.1
tomli==2.0.1
typing_extensions==4.8.0
Werkzeug==2.3.7
boto3==1.29.6
